% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildMetatreeReconciliation.R
\name{BuildMetatreeReconciliation}
\alias{BuildMetatreeReconciliation}
\title{Build metatree taxonomic reconcilaiotn}
\usage{
BuildMetatreeReconciliation(Taxa, SortNames = FALSE)
}
\arguments{
\item{Taxa}{A list of taxa exactly as written in the Paleobiology Database, except with underscores within names instead of spaces.}

\item{SortNames}{Logical indicating whether names should also be sorted (TRUE) or not (FALSE; the default).}
}
\value{
Nothing is returned. Instead the requisite line is sent to the screen ready for copy-pasting into an XML file.
}
\description{
Given a list of one or more names builds metatree taxonomic reconcilaiton line.
}
\details{
Really intended as a tool for the author, this helps automate the process of building XML lines for multiple-taxon reconciliations.
}
\examples{

# Generate a set of two example trees:
Taxa <- c("Elephas_maximus", "Loxodonta_africana", "Loxodonta_cyclotis")

# Convert to MRP and show just matrix:
BuildMetatreeReconciliation(Taxa)

}
\references{
Lloyd, G. T., Bapst, D. W., Friedman, M. and Davis, K. E., 2016. Probabilistic divergence time estimation without branch lengths: dating the origins of dinosaurs, avian flight, and crown birds. \emph{Biology Letters}, \bold{12}, 20160609.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
