% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollapseDuplicateTaxonMRP.R
\name{CollapseDuplicateTaxonMRP}
\alias{CollapseDuplicateTaxonMRP}
\title{Collapses an MRP matrix with duplicate taxa}
\usage{
CollapseDuplicateTaxonMRP(MRPMatrix)
}
\arguments{
\item{MRPMatrix}{An MRP matrix in the format improted by \code{Claddis::ReadMorphNexus}.}
}
\description{
Collapses an MRP matrix with duplicate taxa until all taxa are unique.
}
\details{
Duplicate taxon names are problematic for downstream anslysis. Generally speaking users should ensure all OTU names are unique. However, in some cases it might make more sense to collapse the matrix until all taxa are unique. Typically this will involve reducing the row count (unique taxa means fewer taxa), but increasing the character count (duplicate taxaon means more unique codings)..
}
\examples{

# Nothing yet.

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{NEXUS}
