% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetFilesForClade.R
\name{GetFilesForClade}
\alias{GetFilesForClade}
\title{Gets file list for given clade}
\usage{
GetFilesForClade(HTMLendings)
}
\arguments{
\item{HTMLendings}{Vector of strings giving matrix ending for desired HTML.}
}
\value{
A vector of file names intended for use with the \code{Metatree} function.
}
\description{
Gets file list for given clade from HTML (graemetllouyd.com).
}
\examples{

# Get all ichthyosaur file names
# from graemetlloyd.com:
GetFilesForClade("matricht.html")

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
