% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatree-package.R
\docType{package}
\name{metatree-package}
\alias{metatree-package}
\alias{metatree}
\title{Generating Meta-Analytical Phylogenies}
\description{
A collection of functions to help generate meta-analytical hypotheses of species rrelationships.
}
\examples{

# See individual functions

}
\references{
Lloyd, G. T., Bapst, D. W., Friedman, M. and Davis, K. E., 2016. Probabilistic divergence time estimation without branch lengths: dating the origins of dinosaurs, avian flight, and crown birds. Biology Letters, 12, 20160609.
}
\author{
Graeme T. Lloyd <graemetlloyd@gmail.com>
}
\keyword{metatree,phylogeny,supertree}
