% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Metatree.R
\name{Metatree}
\alias{Metatree}
\title{Builds metatree from source data}
\usage{
Metatree(MRPDirectory, XMLDirectory, TargetClade = "",
  InclusiveDataList = c(), ExclusiveDataList = c(),
  HigherTaxaToCollapse = c(), MissingSpecies = "exclude",
  Interval = NULL, VeilLine = TRUE, SpeciesToExclude = c(),
  IncludeSpecimenLevelOTUs = TRUE, BackboneConstraint = NULL,
  MonophylyConstraint = NULL)
}
\arguments{
\item{MRPDirectory}{The directory in which the MRP files are to be read from.}

\item{XMLDirectory}{The directory in which the XML files are to be read from.}

\item{TargetClade}{The name of the target clade of the metatree (e.g., "Dinosauria").}

\item{InclusiveDataList}{A vector of the data sest to include in the metatree. Can be left empty to just read all files in \code{MRPDirectory} abd \code{XMLDirectory}.}

\item{ExclusiveDataList}{A vector of any data sets to exclude from the metatree. Can be left empty if all data sets in \code{MRPDirectory} abd \code{XMLDirectory} are valid. (Intended to exclude things like oogenera or footprint analyses, other supertree data sets etc.)}

\item{HigherTaxaToCollapse}{Vector of any higher taxa to collapse (e.g., if you are focused on relationships in a stem-group).}

\item{MissingSpecies}{What to do with species assigned to the target clade, but not present in the source data. Options are: "exclude" (excludes these missing species), "genus" (include those species in a genus-level polytomy if the genus is sampled in the source data), and "all"}

\item{Interval}{If restricting the sample to a specific interval of geologic time then use this option (passed to \link{PaleobiologyDBChildFinder} which should be consulted for formatting). Default is NULL (no restriction on ages of tips to be included).}

\item{VeilLine}{A logical indiicating whther to remove older data sets that do not increase taxonomic coverage (TRUE; the default) or not (FALSE). See Lloyd et al. (2016).}

\item{SpeciesToExclude}{Vector of any species to be excluded from the final metatree. E.g., Eshanosaurus, Ricardoestesia.}

\item{IncludeSpecimenLevelOTUs}{A logical indicating whther specimen-level OTUs should (TRUE; the default) or should not (FALSE) be included in the metatree.}

\item{BackboneConstraint}{Either a file name of one of the source data sets or a Newick string to represent a backbone constraint (will enforce topology in final metatree but allows taxa not in topology to fall out inside the constraint). This is not required and the default (NULL) will mean no constraint is applied.}

\item{MonophylyConstraint}{Either a file name of one of the source data sets or a Newick string to represent a monophyly constraint (will enforce topology in final metatree and force taxa not in topology to fall outside the constraint). This is not required and the default (NULL) will mean no constraint is applied.}
}
\value{
TBC.
}
\description{
Builds metatree from source data
}
\details{
Use approach laid out in Lloyd et al. (2008; 2016).
}
\examples{

# Nothing yet.

}
\references{
Lloyd et al. (@016)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
