% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTreeContradiction.R
\name{MultiTreeContradiction}
\alias{MultiTreeContradiction}
\title{Produces tree contradiction distance matrix}
\usage{
MultiTreeContradiction(trees, rescale = TRUE)
}
\arguments{
\item{trees}{An object of class 'multi.phylo'.}

\item{rescale}{Whether or not to rescale contradictions (zero to one).}
}
\value{
A contradiction (distance) matrix.
}
\description{
Given a set of input trees calculates all pariwise tree contradictions.
}
\details{
This is effectively an attempt to do for the paleotree function "treeContradiction" what phytools "multiRF" does for Robinson-Foulds, i.e., operate on all pairwise comparisons instead of a single one.
}
\examples{

# Nothing yet

}
\references{
This contradiction difference measure was introduced in:

Bapst, D. W., H. A. Schreiber, and S. J. Carlson. 2018. Combined Analysis of Extant Rhynchonellida
(Brachiopoda) using Morphological and Molecular Data. \emph{Systematic Biology} 67(1):32-48. doi: 10.1093/sysbio/syx049
}
\seealso{
A less-optimized function for obtaining the contradiction difference measure
is \code{\link[paleotree]{treeContradiction}}, found in package \code{paleotree}.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
