% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTreeDistance.R
\name{MultiTreeDistance}
\alias{MultiTreeDistance}
\title{Multiple tree distances}
\usage{
MultiTreeDistance(trees, distance = "contradiction", rescale = TRUE)
}
\arguments{
\item{trees}{An object of class 'multi.phylo'.}

\item{distance}{The type of disatnce to use, must be one of "RF" (RObinson-Foulds) or "contradiction" (tree contradiction).}

\item{rescale}{Whether or not to rescale the distance (zero to one).}
}
\value{
A symmetric matrix of pairwise tree distances.
}
\description{
Given a set of input trees, calculates all pairwise tree distances.
}
\details{
This function is an incomplete attempt to generalise the \link{MultiTreeContradiction} function to incorporate not just contradiction differences (Bapst et al. 2018), but also the Robinson-Foulds distance metric (Robinson and Foulds 1981).

It is not yet ready for general use.
}
\examples{

# Generate three example trees (with tips labelled A-D):
ExampleTrees <- read.tree(text = c("(A,B,C,D);", "(A,(B,(C,D)));",
  "((A,B),C,D);"))

# Calculate rescaled RF distances matrix:
MultiTreeDistance(ExampleTrees, distance = "RF")

}
\references{
Bapst, D. W., H. A. Schreiber, and S. J. Carlson. 2018. Combined analysis of extant Rhynchonellida (Brachiopoda) using morphological and molecular data. \emph{Systematic Biology}, \bold{67}, 32-48.

Robinson, D. F. and Foulds, L. R., 1981. Comparison of phylogenetic trees. \emph{Mathematical Biosciences}, \bold{53}, 131-147.
}
\seealso{
\code{paleotree::treeContradiction} and \code{phytools::multiRF}.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
