% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleobiologyDBOccurrenceQuerier.R
\name{PaleobiologyDBOccurrenceQuerier}
\alias{PaleobiologyDBOccurrenceQuerier}
\title{Palaeobiology Database Occurrence Querier}
\usage{
PaleobiologyDBOccurrenceQuerier(taxon_nos, original = TRUE,
  breaker = 100, RetainUncertainOccurrences = FALSE)
}
\arguments{
\item{taxon_nos}{A vector of Paleobiology database taxon number(s) to retrieve from the database.}

\item{original}{Whether or not to return the original (TRUE) or resolved version (FALSE) of names.}

\item{breaker}{Size of breaker to use if querying a large number of taxa (reduces load on database of individual queries; default is 100).}

\item{RetainUncertainOccurrences}{Logical indicating whether or not to retain uncertain (i.e., aff, cf., ?, "") occurrences (defaults to FALSE).}
}
\value{
A multi-column matrix with rows as occurrences.
}
\description{
Given a set of Paleobiology Database taxon number(s) returns occurrence information for those tax(a).
}
\details{
Uses the Paleobiology Database (\code{paleobiodb.org}) API to query known taxon numbers and returns information on their occurrence as fossils.
}
\examples{

# Occurrence query for Allosaurus fragilis:
PaleobiologyDBOccurrenceQuerier(taxon_nos = "52962")

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
