% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleobiologyDBTaxaQuerier.R
\name{PaleobiologyDBTaxaQuerier}
\alias{PaleobiologyDBTaxaQuerier}
\title{Palaeobiology Database Taxa Querier}
\usage{
PaleobiologyDBTaxaQuerier(taxon_nos, taxon_names = NULL,
  original = TRUE, interval = NULL, extant = "include",
  stopfororphans = TRUE, breaker = 100)
}
\arguments{
\item{taxon_nos}{A vector of Paleobiology database taxon numbers to retrieve from the database.}

\item{taxon_names}{A vector of taxon names to search for in the database (default left to NULL).}

\item{original}{Whether or not to return the original (TRUE) or resolved version (FALSE).}

\item{interval}{The beginning and ending geologic periods if only wanting taxa from a specified time window (default is NULL).}

\item{extant}{What to do with extant taxa, one of: "only" (only return extant taxa), "exclude" (exclude extant taxa), or "include" (make no distinction, the default).}

\item{stopfororphans}{Whether or not to stop with an Error message for taxa with no parent.}

\item{breaker}{Size of breaker to use if querying a large number of taxa (reduces load on database of individual queries; default is 100).}
}
\value{
An eight-column matrix detailing the original taxon number (if relevant), the valid (resolved) taxon number, the taxon name, the taxon rank (Paleobiology Database rank number), the taxon number of the parent of each taxon, the taxon validity (if relevant; returns NA if already valid), the accepted taxon number (if relevant), and the accepted taxon name (if relevant).
}
\description{
Given a Paleobiology Database taxon number returns basic information on that taxon.
}
\details{
Uses the Paleobiology Database (\code{paleobiodb.org}) API to query known taxon numbers and returns information on their validity, name, and rank. Intended for use in building dynamic taxonomic resolutions when constructing metatree matrices (see Lloyd et al. 2016).
}
\examples{

# Taxon query for Allosaurus fragilis:
PaleobiologyDBTaxaQuerier(taxon_nos = "52962")

}
\references{
Lloyd, G. T., Bapst, D. W., Friedman, M. and Davis, K. E., 2016. Probabilistic divergence time estimation without branch lengths: dating the origins of dinosaurs, avian flight, and crown birds. Biology Letters, 12, 20160609.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
