% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleobiologyDBTaxonQuerier.R
\name{PaleobiologyDBTaxonQuerier}
\alias{PaleobiologyDBTaxonQuerier}
\title{Palaeobiology Database Taxon Querier}
\usage{
PaleobiologyDBTaxonQuerier(taxon_no, taxon_name = NULL, original = TRUE,
  stopfororphans = TRUE, breaker = 100)
}
\arguments{
\item{taxon_no}{A vector of Paleobiology database taxon numbers to retrieve from the database.}

\item{taxon_name}{A vector of taxon names to search for in the database (default left to NULL).}

\item{original}{Whether or not to return the original (TRUE) or resolved version (FALSE).}

\item{stopfororphans}{Whether or not to stop with an Error message if taxon has no parent.}

\item{breaker}{Size of breaker to use if querying a large number of taxa (reduces load on database of individual queries).}
}
\value{
An eight-item list detailing the original taxon number (if relevant), the valid (resolved) taxon number, the taxon name, the taxon rank (Paleobiology Database rank number), the taxon number of the parent of this taxon, the taxon validity (if relevant; returns NA if already valid), the accepted taxon number (if relevant), and the accepted taxon name (if relevant).
}
\description{
Given a Paleobiology Database taxon number returns basic information on that taxon.
}
\details{
Uses the Paleobiology Database (\code{paleobiodb.org}) API to query a known taxon number and returns information on its validity, name, and rank. Intended for use in building dynamic taxonomic resolutions when building metatree matrices (see Lloyd et al. 2016).
}
\examples{

# Taxon query for Allosaurus fragilis:
PaleobiologyDBTaxonQuerier(taxon_no = "52962")

}
\references{
Lloyd, G. T., Bapst, D. W., Friedman, M. and Davis, K. E., 2016. Probabilistic divergence time estimation without branch lengths: dating the origins of dinosaurs, avian flight, and crown birds. Biology Letters, 12, 20160609.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
