% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleoDBWebScraper.R
\name{PaleoDBWebScraper}
\alias{PaleoDBWebScraper}
\title{Scrapes the Paleobiology Database}
\usage{
PaleoDBWebScraper(taxon, taxon.no = NULL, occurrences = FALSE,
  db = "Paleobiodb")
}
\arguments{
\item{taxon}{The taxon name.}

\item{taxon.no}{Alternatively, the taxon number (supercedes taxon name if supplied).}

\item{occurrences}{Whether or not to also return information on occurrences (if available).}

\item{db}{The database to use, either "Paleobiodb" or "Fossilworks".}
}
\value{
A matrix.
}
\description{
Resolve input name with taxonomy in Paleobiology Database
}
\details{
Deprecated as replaced by API functions.
}
\examples{

# Nothing yet
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}

