% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PisaniMRPPrune.R
\name{PisaniMRPPrune}
\alias{PisaniMRPPrune}
\title{Prunes an MRP matrix of redundant characters}
\usage{
PisaniMRPPrune(MRPMatrix)
}
\arguments{
\item{MRPMatrix}{An MRP matrix in the format imported by \code{Claddis::ReadMorphNexus}.}
}
\value{
A modified matrix with any now redundant characters removed.
}
\description{
Prunes an MRP matrix of redundant characters (Pisani et al. 2002)
}
\examples{

# Nothing yet.

}
\references{
Pisani et al. 2002
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
