% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadMetatreeXML.R
\name{ReadMetatreeXML}
\alias{ReadMetatreeXML}
\title{Reads in a metatree-format XML file}
\usage{
ReadMetatreeXML(File, Invisible = TRUE)
}
\arguments{
\item{File}{Path to XML file to read.}

\item{Invisible}{Logical indicating whether to immediately show entire output (FALSE) or not (TRUE, the default).}
}
\value{
A nested list reflecting the nested XML tags of the input file.
}
\description{
Reads in a metatree-format XML file (Lloyd et al. 2016).
}
\examples{

# Nothing yet.

}
\references{
Lloyd, G. T., Bapst, D. W., Friedman, M., and Davis, K. E., 2016. Probabilistic divergence time estimation without branch lengths: dating the origins of dinosaurs, avian flight and crown birds. Biology Letters, 12, 20160609.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
