% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TNTMRP2NEXUSMRP.R
\name{TNTMRP2NEXUSMRP}
\alias{TNTMRP2NEXUSMRP}
\title{Converts TNT MRP output to NEXUS format}
\usage{
TNTMRP2NEXUSMRP(filename)
}
\arguments{
\item{filename}{The TNT input file name.}
}
\value{
A cladistic matrix in \link{ReadMorphNexus} format.
}
\description{
Converts TNT formatted MRP matrix to
}
\details{
TNT (Goloboff et al. 2008) \\#NEXUS (Maddison ????)
}
\examples{

# Nothing yet

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
