% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteMetatreeXML.R
\name{WriteMetatreeXML}
\alias{WriteMetatreeXML}
\title{Writes out a metatree-format XML file}
\usage{
WriteMetatreeXML(XML, File)
}
\arguments{
\item{XML}{The XML file to write. Must be in format imported by \link{ReadMetatreeXML}.}

\item{File}{Path to XML file to write.}
}
\value{
Nothing is returned, but a file is produced.
}
\description{
Writes out a metatree-format XML file.
}
\details{
The metatree XML format (Lloyd et al. 2016) is described in detail in the \link{ReadMetatreeXML} help file and is primarily used by the \link{Metatree} function. However, a user may also wish to edit an XML file in R and then write it to a file, and that is what this function does.
}
\examples{

# Example read line:
#x <- ReadMetatreeXML("Rogers_etal_2012a.xml", Invisible = FALSE)

# Example write line:
#WriteMetatreeXML(x, "Rogers_etal_2012a.xml")

# (Note that this is commented out as it would only work locally,
# but should give the user an idea of the syntax)

}
\references{
Lloyd, G. T., Bapst, D. W., Friedman, M., and Davis, K. E., 2016. Probabilistic divergence time estimation without branch lengths: dating the origins of dinosaurs, avian flight and crown birds. \emph{Biology Letters}, \bold{12}, 20160609.
}
\seealso{
\link{ReadMetatreeXML}.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
