% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollapseDuplicateTaxonMRP.R
\name{CollapseDuplicateTaxonMRP}
\alias{CollapseDuplicateTaxonMRP}
\title{Collapses an MRP matrix with duplicate taxa}
\usage{
CollapseDuplicateTaxonMRP(MRPMatrix)
}
\arguments{
\item{MRPMatrix}{An MRP matrix in the format imported by \code{Claddis::ReadMatrixNEXUS}.}
}
\description{
Collapses an MRP matrix with duplicate taxa until all taxa are unique.
}
\details{
Duplicate taxon names are problematic for downstream analysis. Generally speaking users should ensure all OTU names are unique. However, in some cases it might make more sense to collapse the matrix until all taxa are unique. Typically this will involve reducing the row count (unique taxa means fewer taxa), but increasing the character count (duplicate taxa means more unique codings).

This function automates this process and is used internally by the \link{Metatree} function, but is also made available for individual use here.
}
\examples{

# Build an example matrix:
ExampleMRP <- Claddis::MatrixBuilder(matrix(as.character(c(0, 0,
  0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1)),
  nrow = 6, dimnames = list(LETTERS[1:6], c()), byrow = TRUE))

# Show taxon names are currently unique by inspection:
rownames(ExampleMRP$Matrix_1$Matrix)

# Now rename taxon F as taxon E:
rownames(ExampleMRP$Matrix_1$Matrix)[6] <- "E"

# Show taxon E is duplicated by inspection:
rownames(ExampleMRP$Matrix_1$Matrix)

# Now collapse to just unique taxa using function:
ExampleMRP <- CollapseDuplicateTaxonMRP(ExampleMRP)

# Confirm that matrix has now been collapsed to five unique taxa
# (and four unique characters) by inspection:
ExampleMRP$Matrix_1$Matrix

}
\seealso{
For another means by which MRP matrices require collapsing see \link{PisaniMRPPrune}.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\keyword{MRP,NEXUS}
