% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmbiggenMatrix.R
\name{EmbiggenMatrix}
\alias{EmbiggenMatrix}
\title{Embiggen a cladistic matrix}
\usage{
EmbiggenMatrix(CladisticMatrix, N)
}
\arguments{
\item{CladisticMatrix}{A cladistic matrix in the format imported by \code{Claddis::read_nexus_matrix}.}

\item{N}{The number of times to duplicate the data.}
}
\value{
A cladistic matrix in the format imported by \code{Claddis::read_nexus_matrix}.
}
\description{
Given a cladistic matrix, will duplicate each block of characters N times.
}
\details{
Given a cladistic matrix in the format imported by \code{Claddis::read_nexus_matrix} will duplicate each block N times. Thus if a block contains 10 characters and N is 5 the resulting block will have 50 characters total.

This function exists for internal use in the \link{Metatree} function as a means of upweighting MRP data beyond the limits of TNT (maximum weight 1000) by duplicating blocks. However, it is also made available for individual use here.
}
\examples{

# Build an example matrix of six species and four characters:
ExampleMatrix <- Claddis::build_cladistic_matrix(matrix(as.character(c(0, 0,
  0, 0, 0, 0, 0, 1, 0, 0, 1, 1, 0, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1)),
  nrow = 6, dimnames = list(LETTERS[1:6], c()), byrow = TRUE))

# Inspect starting matrix:
ExampleMatrix$matrix_1$matrix

# Duplicate block five times (20 characters):
ExampleMatrix <- EmbiggenMatrix(ExampleMatrix, 5)

# Inspect matrix to check it now contains 20 characters:
ExampleMatrix$matrix_1$matrix

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
