% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmbiggenMatrix.R
\name{EmbiggenMatrix}
\alias{EmbiggenMatrix}
\title{Embiggen a cladistic matrix}
\usage{
EmbiggenMatrix(CladisticMatrix, N)
}
\arguments{
\item{CladisticMatrix}{A cladistic matrix in the format imported by \code{Claddis::ReadMorphNexus}.}

\item{N}{The number of times to duplicated the data.}
}
\value{
A cladistic matrix in the format imported by \code{Claddis::ReadMorphNexus}.
}
\description{
Given a cladistic matrix duplicates the block N times.
}
\details{
Given a cladistic matrix duplicates the block N times.
}
\examples{

# Nothing yet.

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
