% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListPosition2ListCoordinates.R
\name{ListPosition2ListCoordinates}
\alias{ListPosition2ListCoordinates}
\title{Gives position of the kth item in a list}
\usage{
ListPosition2ListCoordinates(Position, ListDimensions)
}
\arguments{
\item{Position}{The kth value for which the position is desired.}

\item{ListDimensions}{The dimensions (vector length) of each part of the list in order.}
}
\value{
The position, in tersm of both list number ($ListNumber) and list position ($ListPosition), of the kth value.
}
\description{
Gives position of the kth item in a list of vectors.
}
\examples{

# Find the position of the 13th value:.
ListPosition2ListCoordinates(13, c(1,3,2,6,1,3))

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
