% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metatree-package.R
\docType{package}
\name{metatree-package}
\alias{metatree}
\alias{metatree-package}
\title{Generating Meta-Analytical Phylogenies}
\description{
Generates meta-analytical phylogenies.
}
\examples{

# Nothing yet

}
\author{
Graeme T. Lloyd <graemetlloyd@gmail.com>
}
\references{
Lloyd et al. (2008).
}
\keyword{phylogeny,supertree}

