% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MRPCollapse.R
\name{MRPCollapse}
\alias{MRPCollapse}
\title{Collapse MRP matrix}
\usage{
MRPCollapse(matrix)
}
\arguments{
\item{matrix}{An MRP matrix in \link{ReadMorphNexus} format.}
}
\value{
An MRP matrix in \link{ReadMorphNexus} format where all characters are unique.
}
\description{
Collapses an MRP matrix to unique characters
}
\details{
Collapses an MRP matrix to just unique characters followung the protocol aid out in Pisani et al. (2002).
}
\examples{
# Nothing yet.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\references{
Pisani et al. (2002)
}

