% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTreeContradiction.R
\name{MultiTreeContradiction}
\alias{MultiTreeContradiction}
\title{Produces tree contradiction distance matrix}
\usage{
MultiTreeContradiction(trees, rescale = TRUE)
}
\arguments{
\item{trees}{An object of class 'multi.phylo'.}

\item{rescale}{Whether or not to rescale contradictions (zero to one).}
}
\value{
A contradiction (distance) matrix.
}
\description{
Given a set of input trees calculates all pariwise tree contradictions.
}
\details{
This is effectively an attempt to do for the paleotree function "treeContradiction" what phytools "multiRF" does for Robinson-Foulds, i.e., operate on all pairwise comparisons instead of a single one.
}
\examples{

# Nothing yet

}
\references{
Bapst paper (see paleotree packahe and treeContradiction function)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
