% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiTreeDistance.R
\name{MultiTreeDistance}
\alias{MultiTreeDistance}
\title{Pairwise distance matrix for a set of trees}
\usage{
MultiTreeDistance(trees, distance = "contradiction", scale = TRUE)
}
\arguments{
\item{trees}{An object of class 'multi.phylo'.}

\item{distance}{The type of disatnce to use, must be one of "RF" (RObinson-Foulds) or "contradiction" (tree contradiction).}

\item{scale}{Whether or not to scale the distance (zero to one).}
}
\value{
A pairwise tree distance matrix.
}
\description{
Given a set of input trees calculates all pariwise tree distances.
}
\details{
This is effectively an attempt to both do a multi-tree version of the paleotree function "treeContradiction" and a faster version of phytools' "multiRF" function when dealing with very large number of trees or tips (hundreds of trees or tips).
The function works by converting the partitions into matrices and operating on those rather than the tree objects.
This can substantially reduce the memory required and hence sped up the calulcations, but only when data sets are large.
}
\examples{

# Nothing yet

}
\references{
Bapst paper (see paleotree package and treeContradiction function)
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
