% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleobiologyDBChildFinder.R
\name{PaleobiologyDBChildFinder}
\alias{PaleobiologyDBChildFinder}
\title{Palaeobiology Database Child Finder}
\usage{
PaleobiologyDBChildFinder(taxon_nos, taxon_names = NULL,
  original = TRUE, interval = NULL, extant = "include",
  validonly = TRUE, returnrank = NULL, breaker = 100)
}
\arguments{
\item{taxon_nos}{The Paleobiology database taxon number.}

\item{taxon_names}{A taxon name to search for in the database (default left to NULL); overrides taxon_nos if used.}

\item{original}{Whether or not to return the original (TRUE) or resolved version (FALSE).}

\item{interval}{The beginning and ending geologic periods if only wanting taxa from a specified time window (default is NULL).}

\item{extant}{What to do with extant taxa, one of: "only" (only return extant taxa), "exclude" (exclude extant taxa), or "include" (make no distinction, the default).}

\item{validonly}{Whether or not to only retunr valid taxa (TRUE) or all taxa (FALSE).}

\item{returnrank}{Whether or not to only return taxa of a specific rank (e.g., "3" for species, "5" for genera). See Paleobiology Database API for more infomation.}

\item{breaker}{Size of breaker to use if querying a large number of taxa (reduces load on database of individual queries; default is 100).}
}
\value{
A ten-column matrix detailing the original taxon number (if relevant), the valid (resolved) taxon number, the taxon name, the taxon rank (Paleobiology Database rank number), the taxon number of the parent of this taxon, the taxon validity (if relevant; returns NA if already valid), the accepted taxon number (if relevant), the accepted taxon name (if relevant) of all species-level children found, the attribution of the original name as currently entered in the database, and whether ("1") or not ("0") the species is extant.
}
\description{
Given a Paleobiology Database taxon number returns basic information on all species-level children.
}
\details{
Uses the Paleobiology Database (\code{paleobiodb.org}) API to query a known taxon number (or name) and returns information on the validity, name, and rank of all its species-level children. Intended for use in building dynamic taxonomic resolutions when building metatree matrices (see Lloyd et al. 2016).
}
\examples{

# Taxon query for Loxommatinae:
PaleobiologyDBChildFinder("339413", returnrank = "3")

}
\references{
Lloyd, G. T., Bapst, D. W., Friedman, M. and Davis, K. E., 2016. Probabilistic divergence time estimation without branch lengths: dating the origins of dinosaurs, avian flight, and crown birds. Biology Letters, 12, 20160609.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
