% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PaleobiologyDBTaxonQuerier.R
\name{PaleobiologyDBTaxonQuerier}
\alias{PaleobiologyDBTaxonQuerier}
\title{Palaeobiology Database Taxon Querier}
\usage{
PaleobiologyDBTaxonQuerier(taxon_no, original = TRUE)
}
\arguments{
\item{taxon_no}{The Paleobiology database taxon number.}

\item{original}{Whether or not to return the original (TRUE) or resolved version (FALSE).}
}
\value{
A six-item list detailing the original taxon number (if relevant), the valid (resolved) taxon number, the taxon name, the taxon rank (Paleobiology Database rank number), the taxon number of the parent of this taxon, and the taxon validity (if relevant; returns NA if already valid).
}
\description{
Given a Paleobiology Database taxon number returns basic information on that taxon.
}
\details{
Uses the Paleobiology Database (\code{paleobiodb.org}) API to query a known taxon number and returns information on its validity, name, and rank. Intended for use in building dynamic taxonomic resolutions when building metatree matrices (see Lloyd et al. 2016).
}
\examples{

# Taxon query for Allosaurus fragilis:
PaleobiologyDBTaxonQuerier("52962")

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
\references{
Lloyd, G. T., Bapst, D. W., Friedman, M. and Davis, K. E., 2016. Probabilistic divergence time estimation without branch lengths: dating the origins of dinosaurs, avian flight, and crown birds. Biology Letters, 12, 20160609.
}

