% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvivingLineages.R
\name{SurvivingLineages}
\alias{SurvivingLineages}
\title{Finds unique surviving lineages of an extinction}
\usage{
SurvivingLineages(Tree, SurvivorVictim)
}
\arguments{
\item{Tree}{A single tree in ape's phylo format.}

\item{SurvivorVictim}{A numeric vector of victims (0) and survivors (1) with names matching to Tree$tip.label.}
}
\value{
A list of surviving lineages (may be clades - vectors of length greater than one) or individual tips (vectors of length one).
}
\description{
Finds unique surviving lineages of an extinction given a tree and vector of binary (extinction or survival) values.
}
\details{
Given a tree that crosses some extinction boundary, with both extinct and surviving species as tips, we might be interested in knowing how many individual lineages survived. This function finds unique lineages (sets of tips) that survive the boundary.
}
\examples{

# Make sure random examples work:
set.seed(17)

# Generate a random 10 tip tree:
Tree <- ape::rtree(10)

# Remove branch lengths (not used here):
Tree$edge.length <- NULL

# Create random vector of survivors (!) and victims (0):
SurvivorVictim <- sample(c(0, 1), size = 10, replace = TRUE)

# Assign tip names to survivors and victims:
names(SurvivorVictim) <- Tree$tip.label

# Run surviving lineages function on data:
SurvivingLineages(Tree, SurvivorVictim)

# Visually confirm results:
plot(Tree)
ape::tiplabels(SurvivorVictim)

# Generate multiple trees with same tip names:
Trees <- ape::rmtree(10, 10)

# Calculate just total number of survivors on each tree:
unlist(lapply(Trees, function(x) length(SurvivingLineages(x, SurvivorVictim = SurvivorVictim))))

}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
