% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteMetatreeXML.R
\name{WriteMetatreeXML}
\alias{WriteMetatreeXML}
\title{Writes a metatree-format XML file}
\usage{
WriteMetatreeXML(XML, File)
}
\arguments{
\item{XML}{The XML file to write. Must be in format imported by \link{ReadMetatreeXML}.}

\item{File}{Path to XML file to write.}
}
\value{
A nested list reflecting the nested XML tags of the input file.
}
\description{
Writes out a metatree-format XML file (Lloyd et al. 2016).
}
\examples{

# Nothing yet.

}
\references{
Lloyd, G. T., Bapst, D. W., Friedman, M., and Davis, K. E., 2016. Probabilistic divergence time estimation without branch lengths: dating the origins of dinosaurs, avian flight and crown birds. Biology Letters, 12, 20160609.
}
\author{
Graeme T. Lloyd \email{graemetlloyd@gmail.com}
}
